Changed how the jacks are handled in the Euchre game rules.
